+++ lib/td_dd/data_structures/data_structures.ex
@@ -34,7 +34,7 @@ defmodule TdDd.DataStructures do
   # Data structure version associations preloaded for some views
   @preload_dsv_assocs [:classifications, data_structure: :system]
 
-  def list_data_structures(clauses \\ %{}, options \\ []) do
+  def list_data_structures(clauses \\ %{}, opts \\ []) do
     clauses
     |> Enum.reduce(DataStructure, fn
       {:external_id, external_ids}, q when is_list(external_ids) ->
@@ -58,7 +58,7 @@ defmodule TdDd.DataStructures do
     |> select_merge([_, sn], %{latest_note: sn.df_content})
     |> preload(:system)
     |> Repo.all()
-    |> Enum.map(&enrich(&1, options))
+    |> Enum.map(&enrich(&1, opts))
   end
 
   def list_data_structure_versions(clauses \\ %{}) do
@@ -95,18 +95,18 @@ defmodule TdDd.DataStructures do
     enriched_structure_version!(id)
   end
 
-  def get_data_structure_version!(data_structure_version_id, options) do
+  def get_data_structure_version!(data_structure_version_id, opts) do
     data_structure_version_id
     |> get_data_structure_version!()
-    |> enrich(options)
+    |> enrich(opts)
   end
 
-  def get_data_structure_version!(data_structure_id, version, options) do
+  def get_data_structure_version!(data_structure_id, version, opts) do
     DataStructureVersion
     |> Repo.get_by!(data_structure_id: data_structure_id, version: version)
     |> Map.get(:id)
     |> enriched_structure_version!(preload: @preload_dsv_assocs)
-    |> enrich(options)
+    |> enrich(opts)
   end
 
   def get_cached_content(%{} = content, %{data_structure_type: %{template_id: template_id}}) do
@@ -125,75 +125,75 @@ defmodule TdDd.DataStructures do
 
   defp enrich(target, nil = _opts), do: target
 
-  defp enrich(%DataStructure{} = ds, options) do
+  defp enrich(%DataStructure{} = ds, opts) do
     ds
     |> StructureEnricher.enrich()
-    |> enrich(options, :versions, &Repo.preload(&1, :versions))
-    |> enrich(options, :latest, &get_latest_version/1)
-    |> enrich(options, :source, &get_source/1)
-    |> enrich(options, :metadata_versions, &get_metadata_versions/1)
+    |> enrich(opts, :versions, &Repo.preload(&1, :versions))
+    |> enrich(opts, :latest, &get_latest_version/1)
+    |> enrich(opts, :source, &get_source/1)
+    |> enrich(opts, :metadata_versions, &get_metadata_versions/1)
   end
 
   defp enrich(%DataStructureVersion{id: id} = _data_structure_version, :defaults) do
     enriched_structure_version!(id, preload: [data_structure: :source])
   end
 
-  defp enrich(%DataStructureVersion{} = dsv, options) do
+  defp enrich(%DataStructureVersion{} = dsv, opts) do
     deleted = not is_nil(Map.get(dsv, :deleted_at))
-    with_confidential = Enum.member?(options, :with_confidential)
+    with_confidential = Enum.member?(opts, :with_confidential)
 
     dsv
     |> enrich(:defaults)
-    |> enrich(options, :classifications, &get_classifications/1)
-    |> enrich(options, :system, &get_system/1)
+    |> enrich(opts, :classifications, &get_classifications/1)
+    |> enrich(opts, :system, &get_system/1)
     |> enrich(
-      options,
+      opts,
       :parents,
       &get_parents(&1, deleted: deleted, with_confidential: with_confidential)
     )
     |> enrich(
-      options,
+      opts,
       :children,
       &get_children(&1, deleted: deleted, with_confidential: with_confidential)
     )
     |> enrich(
-      options,
+      opts,
       :siblings,
       &get_siblings(&1, deleted: deleted, with_confidential: with_confidential)
     )
     |> enrich(
-      options,
+      opts,
       :data_fields,
       &get_field_structures(&1,
         deleted: deleted,
-        preload: if(Enum.member?(options, :profile), do: [data_structure: :profile], else: []),
+        preload: if(Enum.member?(opts, :profile), do: [data_structure: :profile], else: []),
         with_confidential: with_confidential
       )
     )
-    |> enrich(options, :data_field_degree, &get_field_degree/1)
-    |> enrich(options, :data_field_links, &get_field_links/1)
+    |> enrich(opts, :data_field_degree, &get_field_degree/1)
+    |> enrich(opts, :data_field_links, &get_field_links/1)
     |> enrich(
-      options,
+      opts,
       :relations,
       &get_relations(&1, deleted: deleted, default: false, with_confidential: with_confidential)
     )
-    |> enrich(options, :relation_links, &get_relation_links/1)
-    |> enrich(options, :versions, &get_versions/1)
-    |> enrich(options, :degree, &get_degree/1)
-    |> enrich(options, :profile, &get_profile/1)
-    |> enrich(options, :links, &get_structure_links/1)
-    |> enrich(options, :source, &get_source/1)
-    |> enrich(options, :metadata_versions, &get_metadata_versions/1)
-    |> enrich(options, :data_structure_type, &get_data_structure_type/1)
-    |> enrich(options, :tags, &get_tags/1)
-    |> enrich(options, :grants, &get_grants/1)
-    |> enrich(options, :grant, &get_grant(&1, options[:user_id]))
+    |> enrich(opts, :relation_links, &get_relation_links/1)
+    |> enrich(opts, :versions, &get_versions/1)
+    |> enrich(opts, :degree, &get_degree/1)
+    |> enrich(opts, :profile, &get_profile/1)
+    |> enrich(opts, :links, &get_structure_links/1)
+    |> enrich(opts, :source, &get_source/1)
+    |> enrich(opts, :metadata_versions, &get_metadata_versions/1)
+    |> enrich(opts, :data_structure_type, &get_data_structure_type/1)
+    |> enrich(opts, :tags, &get_tags/1)
+    |> enrich(opts, :grants, &get_grants/1)
+    |> enrich(opts, :grant, &get_grant(&1, opts[:user_id]))
   end
 
-  defp enrich(%{} = target, options, key, fun) do
+  defp enrich(%{} = target, opts, key, fun) do
     target_key = get_target_key(key)
 
-    case Enum.member?(options, key) do
+    case Enum.member?(opts, key) do
       false -> target
       true -> Map.put(target, target_key, fun.(target))
     end
@@ -230,37 +230,37 @@ defmodule TdDd.DataStructures do
     |> Repo.one()
   end
 
-  def get_field_structures(data_structure_version, options) do
+  def get_field_structures(data_structure_version, opts) do
     data_structure_version
     |> Ecto.assoc(:children)
     |> where([child], child.class == "field")
     |> join(:inner, [child], child in assoc(child, :data_structure))
     |> with_confidential(
-      Keyword.get(options, :with_confidential),
+      Keyword.get(opts, :with_confidential),
       dynamic([_child, _parent, child_ds], child_ds.confidential == false)
     )
-    |> with_deleted(options, dynamic([child], is_nil(child.deleted_at)))
+    |> with_deleted(opts, dynamic([child], is_nil(child.deleted_at)))
     |> select([child], child)
     |> Repo.all()
-    |> Repo.preload(options[:preload] || [])
+    |> Repo.preload(opts[:preload] || [])
   end
 
-  def get_children(%DataStructureVersion{id: id}, options \\ []) do
+  def get_children(%DataStructureVersion{id: id}, opts \\ []) do
     DataStructureRelation
     |> where([r], r.parent_id == ^id)
     |> join(:inner, [r], parent in assoc(r, :child))
     |> join(:inner, [r, _parent, _], r in assoc(r, :relation_type))
     |> join(:inner, [_r, child, _], child in assoc(child, :data_structure))
     |> with_deleted(
-      Keyword.get(options, :deleted),
+      Keyword.get(opts, :deleted),
       dynamic([_, parent, _], is_nil(parent.deleted_at))
     )
     |> with_confidential(
-      Keyword.get(options, :with_confidential),
+      Keyword.get(opts, :with_confidential),
       dynamic([_, _child, _, ds_child], ds_child.confidential == false)
     )
     |> relation_type_condition(
-      Keyword.get(options, :default),
+      Keyword.get(opts, :default),
       dynamic([_, _child, relation_type, _], relation_type.name == "default"),
       dynamic([_, _child, relation_type, _], relation_type.name != "default")
     )
@@ -272,26 +272,26 @@ defmodule TdDd.DataStructures do
       relation_type: relation_type
     })
     |> Repo.all()
-    |> select_structures(Keyword.get(options, :default))
+    |> select_structures(Keyword.get(opts, :default))
   end
 
-  def get_parents(%DataStructureVersion{id: id}, options \\ []) do
+  def get_parents(%DataStructureVersion{id: id}, opts \\ []) do
     DataStructureRelation
     |> where([r], r.child_id == ^id)
     |> join(:inner, [r], parent in assoc(r, :parent))
     |> join(:inner, [r, _parent], relation_type in assoc(r, :relation_type))
     |> join(:inner, [_r, parent, _relation_type], parent in assoc(parent, :data_structure))
     |> with_deleted(
-      Keyword.get(options, :deleted),
+      Keyword.get(opts, :deleted),
       dynamic([_, parent, _, _], is_nil(parent.deleted_at))
     )
     |> relation_type_condition(
-      Keyword.get(options, :default),
+      Keyword.get(opts, :default),
       dynamic([_, _parent, relation_type, _], relation_type.name == "default"),
       dynamic([_, _parent, relation_type, _], relation_type.name != "default")
     )
     |> with_confidential(
-      Keyword.get(options, :with_confidential),
+      Keyword.get(opts, :with_confidential),
       dynamic([_, _parent, _relation_type, parent_ds], parent_ds.confidential == false)
     )
     |> order_by([_, parent, _, _], asc: parent.data_structure_id, desc: parent.version)
@@ -302,10 +302,10 @@ defmodule TdDd.DataStructures do
       relation_type: relation_type
     })
     |> Repo.all()
-    |> select_structures(Keyword.get(options, :default))
+    |> select_structures(Keyword.get(opts, :default))
   end
 
-  def get_siblings(%DataStructureVersion{id: id}, options \\ []) do
+  def get_siblings(%DataStructureVersion{id: id}, opts \\ []) do
     DataStructureRelation
     |> where([r], r.child_id == ^id)
     |> join(:inner, [r], parent in assoc(r, :parent))
@@ -321,28 +321,28 @@ defmodule TdDd.DataStructures do
       sibling in assoc(sibling, :data_structure)
     )
     |> with_deleted(
-      options,
+      opts,
       dynamic(
         [_r, parent, _parent_rt, _r_c, _child_rt, _sibling, _sibling_ds],
         is_nil(parent.deleted_at)
       )
     )
     |> with_deleted(
-      options,
+      opts,
       dynamic(
         [_r, parent, _parent_rt, _r_c, _child_rt, sibling, _sibling_ds],
         is_nil(sibling.deleted_at)
       )
     )
     |> with_confidential(
-      Keyword.get(options, :with_confidential),
+      Keyword.get(opts, :with_confidential),
       dynamic(
         [_r, _parent, _parent_rt, _r_c, _child_rt, _sibling, sibling_ds],
         sibling_ds.confidential == false
       )
     )
     |> relation_type_condition(
-      Keyword.get(options, :default),
+      Keyword.get(opts, :default),
       dynamic(
         [_r, _parent, parent_rt, _r_c, _child_rt, _sibling, _sibling_ds],
         parent_rt.name == "default"
@@ -353,7 +353,7 @@ defmodule TdDd.DataStructures do
       )
     )
     |> relation_type_condition(
-      Keyword.get(options, :default),
+      Keyword.get(opts, :default),
       dynamic(
         [_r, _parent, _parent_rt, _r_c, child_rt, _sibling, _sibling_ds],
         child_rt.name == "default"
@@ -374,9 +374,9 @@ defmodule TdDd.DataStructures do
     |> Enum.uniq_by(& &1.data_structure_id)
   end
 
-  defp get_relations(%DataStructureVersion{} = version, options) do
-    parents = get_parents(version, options)
-    children = get_children(version, options)
+  defp get_relations(%DataStructureVersion{} = version, opts) do
+    parents = get_parents(version, opts)
+    children = get_children(version, opts)
 
     %{parents: parents, children: children}
   end
@@ -647,22 +647,22 @@ defmodule TdDd.DataStructures do
      }}
   end
 
-  def get_latest_version(target, options \\ [])
+  def get_latest_version(target, opts \\ [])
 
   def get_latest_version(nil, _), do: nil
 
-  def get_latest_version(%DataStructure{id: id}, options) do
-    get_latest_version(id, options)
+  def get_latest_version(%DataStructure{id: id}, opts) do
+    get_latest_version(id, opts)
   end
 
-  def get_latest_version(data_structure_id, options) do
+  def get_latest_version(data_structure_id, opts) do
     DataStructureVersion
     |> where(data_structure_id: ^data_structure_id)
     |> distinct(:data_structure_id)
     |> order_by(desc: :version)
     |> preload(data_structure: :source)
     |> Repo.one()
-    |> enrich(options)
+    |> enrich(opts)
   end
 
   def put_domain_id(data, %{} = domain_map, domain) when is_binary(domain) do
@@ -775,16 +775,16 @@ defmodule TdDd.DataStructures do
     end
   end
 
-  def get_latest_version_by_external_id(external_id, options \\ []) do
+  def get_latest_version_by_external_id(external_id, opts \\ []) do
     DataStructureVersion
-    |> with_deleted(options, dynamic([dsv], is_nil(dsv.deleted_at)))
+    |> with_deleted(opts, dynamic([dsv], is_nil(dsv.deleted_at)))
     |> distinct(:data_structure_id)
     |> order_by(desc: :version)
     |> join(:inner, [data_structure], ds in assoc(data_structure, :data_structure))
     |> where([_, ds], ds.external_id == ^external_id)
     |> select_merge([_, ds], %{external_id: ds.external_id})
     |> Repo.one()
-    |> enrich(options[:enrich])
+    |> enrich(opts[:enrich])
   end
 
   @spec get_latest_versions([non_neg_integer]) :: [DataStructureVersion.t()]
@@ -796,8 +796,8 @@ defmodule TdDd.DataStructures do
     |> Repo.all()
   end
 
-  defp with_deleted(query, options, dynamic) when is_list(options) do
-    include_deleted = Keyword.get(options, :deleted, true)
+  defp with_deleted(query, opts, dynamic) when is_list(opts) do
+    include_deleted = Keyword.get(opts, :deleted, true)
     with_deleted(query, include_deleted, dynamic)
   end
 
@@ -934,16 +934,16 @@ defmodule TdDd.DataStructures do
 
   defp do_profile_source(dsv, _source), do: dsv
 
-  def list_data_structure_tags(options \\ []) do
+  def list_data_structure_tags(opts \\ []) do
     DataStructureTag
     |> Repo.all()
-    |> Repo.preload(options[:preload] || [])
+    |> Repo.preload(opts[:preload] || [])
   end
 
-  def get_data_structure_tag!(id, options \\ []) do
+  def get_data_structure_tag!(id, opts \\ []) do
     DataStructureTag
     |> Repo.get!(id)
-    |> Repo.preload(options[:preload] || [])
+    |> Repo.preload(opts[:preload] || [])
   end
 
   def create_data_structure_tag(attrs \\ %{}) do