+++ CHANGELOG.md
@@ -6,6 +6,13 @@
 
 - [TD-3484] GraphQL API on `/api/v2`
 
+### Changed
+
+- [TD-3039] `GET /api/data_structures/:data_structure_id/versions/:id` now
+  includes mutable metadata in the `metadata` field. The `metadata_versions`
+  field is no longer included in the response. The `metadata_fields` field in
+  the data structure type responses now include mutable metadata fields.
+
 ## [4.26.0] 2021-08-16
 
 ### Added